refactor: renamed local simple tables for mapman and interpro on gene show page
