+++ components/tables/generics/LocalSimpleTable.tsx
